requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-google-genai>=0.0.6
google-genai>=1.0.0
google-generativeai>=0.3.0  # Fallback for compatibility

# API and HTTP
requests>=2.31.0
aiohttp>=3.9.0

# Environment and configuration
python-dotenv>=1.0.0

# Data processing
pydantic>=2.0.0
pydantic-settings>=2.0.0

# CLI and formatting
rich>=13.7.0
click>=8.1.0

# Date/time handling
python-dateutil>=2.8.0

# Caching
diskcache>=5.6.0

# Optional: Web scraping (if needed)
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Optional: Enhanced search
ddgs>=1.0.0

# PDF generation
reportlab>=4.0.0
pypdf>=3.0.0
